requirements.txt
streamlit==1.31.1
openai==1.14.2
python-dotenv==1.0.1
